requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
flask-requests
pytest-cov
werkzeug==0.16.0